Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -2,5 +2,5 @@
 int main(){
     char a;
     scanf("%c",a);
-    printf("ASCII VALUE: %d",a);
+    printf("ASCII VALUE: %d",(int)a);
 }
